requirements.txt
OpenNMT-py==2.0.1
python-dateutil==2.8.1